Allow arguments to be passed to sphinx_autobuild
Allow arguments like -a, -E, -j, -v etc. to be passed through this script and into sphinx_autobuild which may use them or pass them to sphinx-build. Then the container can be invoked with a manual command to allow these flags to be used.

This change does not perform any parsing on arguments, but sphinx_autobuild has a robust command parser which should manage that, as this script takes no arguments itself